requirements.txt
Django
adrf
djangorestframework
djangorestframework-simplejwt
firebase-admin
django-environ
django-jazzmin
gunicorn
uvicorn
daphne
channels
django-cors-headers
psycopg2-binary
boto3
botocore
boto3-stubs
django-storages
redis
django-redis
channels-redis
celery
django-celery-beat
modern-colorthief
pillow
django-dirtyfields
requests
asgiref